+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This CHANGELOG, TERMS, and CONTRIBUTING documentation to support open sourcing this project (representabot#17).
 - Basic unit tests for data.py — it's a start! (representabot#19).
 - Pre-commit hook to run unit tests.
+- Command line arguments for "congress" and "session"
 
 ### Changed
 - Made maximum number of tweets per run configurable with `MAX_TWEETS` environment variable.